Can now move the selected item
Move the selected item by calling the function in class example. Right
now i am calling the function in the controller under service mouse
move, this needs be changed to only work on mouse drag, not any mouse
move

Right now if you rotate somthing you wont be able to click on it or move
it correctly. so don't try it